Auto-create .git/hooks/pre-commit on ./configure
The hook created checks to see if dev/tools/pre-commit exists, and, if
so, runs it.  This way, we don't have to do any fancy logic to update
the git pre-commit hook.  The configure script never overwrites an
existing precommit hook, so users can disable it by creating an empty
pre-commit hook.

The check for existence is so that if users check out an old version of
Coq, attempting to commit won't give an error about non-existent files.